refactor: json config
